[safety-rules] bring the client related code back to consensus
So that I can use prometheus counters.
