+++ consensus/Cargo.toml
@@ -38,9 +38,11 @@ executor-types = { path = "../execution/executor-types", version = "0.1.0" }
 lcs = { path = "../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-config = { path = "../config", version = "0.1.0" }
 libra-crypto = { path = "../crypto/crypto", version = "0.1.0" }
+libra-global-constants = { path = "../config/global-constants", version = "0.1.0"}
 libra-logger = { path = "../common/logger", version = "0.1.0" }
 libra-mempool = { path = "../mempool", version = "0.1.0" }
 libra-metrics = { path = "../common/metrics", version = "0.1.0" }
+libra-secure-net = { path = "../secure/net", version = "0.1.0" }
 libra-secure-storage = { path = "../secure/storage", version = "0.1.0" }
 libra-security-logger = { path = "../common/security-logger", version = "0.1.0" }
 libra-temppath = { path = "../common/temppath", version = "0.1.0" }
@@ -60,7 +62,13 @@ tempfile = "3.1.0"
 
 vm-genesis = { path = "../language/tools/vm-genesis", version = "0.1.0" }
 vm-validator = { path = "../vm-validator", version = "0.1.0" }
+workspace-builder = { path = "../common/workspace-builder", version = "0.1.0" }
 
 [features]
 default = []
 fuzzing = ["proptest", "consensus-types/fuzzing", "libra-config/fuzzing", "libra-crypto/fuzzing", "libra-mempool/fuzzing", "libra-types/fuzzing", "safety-rules/testing"]
+
+[[bench]]
+name = "safety_rules"
+harness = false
+path = "src/safety_rules_client/benches/safety_rules.rs"
