+++ consensus/safety-rules/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0.31"
 once_cell = "1.4.0"
-rand = { version = "0.7.3", default-features = false }
 
 consensus-types = { path = "../consensus-types", version = "0.1.0" }
 lcs = { path = "../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
@@ -20,20 +19,9 @@ libra-logger = { path = "../../common/logger", version = "0.1.0" }
 libra-secure-net = { path = "../../secure/net", version = "0.1.0" }
 libra-secure-push-metrics = { path = "../../secure/push-metrics", version = "0.1.0" }
 libra-secure-storage = { path = "../../secure/storage", version = "0.1.0" }
-libra-temppath = { path = "../../common/temppath", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
-libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 serde = { version = "1.0.112", default-features = false }
 thiserror = "1.0.20"
-
-[dev-dependencies]
-criterion = "0.3.2"
-tempfile = "3.1.0"
-workspace-builder = { path = "../../common/workspace-builder", version = "0.1.0" }
-
-[[bench]]
-name = "safety_rules"
-harness = false
 
 [features]
 default = []
