+++ consensus/src/lib.rs
@@ -22,6 +22,7 @@ mod network;
 mod network_tests;
 mod persistent_liveness_storage;
 mod round_manager;
+mod safety_rules_client;
 mod state_computer;
 mod state_replication;
 #[cfg(any(test, feature = "fuzzing"))]
