+++ consensus/src/safety_rules_client/spawned_process.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::remote_service::RemoteService;
+use safety_rules::RemoteService;
 
 use libra_config::config::{NodeConfig, PersistableConfig, SafetyRulesService};
 use libra_temppath::TempPath;
