+++ consensus/src/safety_rules_client/tests/serializer.rs
@@ -1,8 +1,9 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{test_utils, tests::suite, SafetyRulesManager, TSafetyRules};
+use crate::safety_rules_client::{test_utils, tests::suite, SafetyRulesManager};
 use libra_types::validator_signer::ValidatorSigner;
+use safety_rules::TSafetyRules;
 
 #[test]
 fn test() {
